removed tmp_prepare + flake8
